+++ hydra-node/src/Hydra/Network/Reliability.hs
@@ -1,9 +1,20 @@
 {-# OPTIONS_GHC -Wno-orphans #-}
 
--- | A `Network` layer that guarantees delivery of `msg` in order even in the
--- face of transient connection failures.
+-- | A `Network` layer that provides resilience in the face of network
+-- connectivity loss and (most) crashes.
 --
--- This layer implements an algorithm based on _vector clocks_, loosely inspired from
+-- This network layer takes care of 2 aspects that together improve the
+-- reliability and operability of a Hydra cluster, making it more tolerant to
+-- transient failures and therefore alleviating the need to prematurely close
+-- heads:
+--
+-- 1. To safeguard against lost of connectivity, it keeps track of messages
+-- indices and resend lost messages,
+-- 2. To safeguard against crashes, it persists all sent messages and indices on disk.
+--
+-- == Messages ordering & resending
+--
+-- This layer implements an algorithm based on /vector clocks/, loosely inspired from
 -- /Reliable consistent broadcast/ algorithms with FIFO ordering as presented in
 -- [Introduction to Reliable and Secure Distributed
 -- Programming](https://www.distributedprogramming.net), ch. 3.9, by Cachin et al.
@@ -22,7 +33,7 @@
 --       * It is discarded if the index for the sender's party in the message is
 --         not exactly one more than the latest known index,
 --
---       * If our own party's index as broadcasted by the sender is lower than
+--       * If our own party's index, as broadcasted by the sender, is lower than
 --         our latest known index, and the peer appears /quiescent/ we resend
 --         all the "missing" messages.
 --
@@ -45,17 +56,28 @@
 --     counter but only take into account their view of /our/ counter in order
 --     to compute whether or not to resend previous messages.
 --
--- NOTE: This layer does not guarantee resilience in the crash-recovery setting,
--- eg. if a process crashes and then later recovers. It may work because `Ping`s
--- will trigger some resending to the peer which starts from scratch. To provide
--- this guarantee in full, we should add /logging/ capability that persist sent and
--- received messages before communicating with the applicative layer.
+-- == Messages persistence
 --
--- NOTE: Messages sent are /currently/ kept indefinitely in memory which, in the
--- case of very long-running head with a high frequency of transaction
--- submission, can lead to memory growth. In practice, this can be mitigated by
--- closing and reopening such heads, but it must be addressed in a not too
--- distant future, eg. perhaps when implementing persistence of messages.
+-- The `MessagePersistence` handle defines an interface that's used to load and
+-- persist both the sequence of messages `broadcast`ed, and the /vector clock/
+-- representing our current view of the all peers' knowledge about
+-- messages. Because it's hard to guarantee atomicity of IO operations on files,
+-- we only save the indices when we /broadcast/ and use a local cache in the
+-- `NetworkCallback`: The underlying layers can callback us concurrently,
+-- eg. when each connection to peers is managed by a dedicated thread.
+--
+-- __NOTE__: We do not recover (at least) from one particular crash situation: When
+-- the `withReliability` "delivers" a message to the calling code, it's usually
+-- put in a queue wrapped into a `NetworkEvent` and then handled later
+-- on. Should the node crashes at that particular point, it won't be resubmitted
+-- the same message later because the message indices could have been updated
+-- and written to on disk already.
+--
+-- __NOTE__: Messages sent are /currently/ kept indefinitely on disk because we
+-- don't set any bound to how far from the past a peer would need to be resent
+-- messages. In the case of very long-running head with a high frequency of
+-- transaction submission, can lead to significant storage use. Should this
+-- become a problem, this can be mitigated by closing and reopening a head.
 module Hydra.Network.Reliability where
 
 import Hydra.Prelude hiding (empty, fromList, length, replicate, zipWith)
@@ -63,10 +85,10 @@ import Hydra.Prelude hiding (empty, fromList, length, replicate, zipWith)
 import Cardano.Binary (serialize')
 import Cardano.Crypto.Util (SignableRepresentation (getSignableRepresentation))
 import Control.Concurrent.Class.MonadSTM (
-  MonadSTM (readTQueue, readTVarIO, writeTQueue),
-  modifyTVar',
+  MonadSTM (readTQueue, writeTQueue),
   newTQueueIO,
   newTVarIO,
+  readTVarIO,
   writeTVar,
  )
 import Control.Tracer (Tracer)
@@ -86,6 +108,7 @@ import Hydra.Network (Network (..), NetworkComponent)
 import Hydra.Network.Authenticate (Authenticated (..))
 import Hydra.Network.Heartbeat (Heartbeat (..), isPing)
 import Hydra.Party (Party)
+import Hydra.Persistence (Persistence (..), PersistenceIncremental (..))
 import Test.QuickCheck (getPositive, listOf)
 
 data ReliableMsg msg = ReliableMsg
@@ -107,14 +130,28 @@ instance (FromCBOR msg) => FromCBOR (ReliableMsg msg) where
 instance ToCBOR msg => SignableRepresentation (ReliableMsg msg) where
   getSignableRepresentation = serialize'
 
+-- | Log entries specific to this network layer.
+--
+-- __NOTE__: Log items are documented in a YAML schema file which is not
+-- currently public, but should be.
 data ReliabilityLog
   = Resending {missing :: Vector Int, acknowledged :: Vector Int, localCounter :: Vector Int, partyIndex :: Int}
   | BroadcastCounter {partyIndex :: Int, localCounter :: Vector Int}
   | BroadcastPing {partyIndex :: Int, localCounter :: Vector Int}
   | Received {acknowledged :: Vector Int, localCounter :: Vector Int, partyIndex :: Int}
   | Ignored {acknowledged :: Vector Int, localCounter :: Vector Int, partyIndex :: Int}
-  | ReliabilityFailedToFindMsg {failedToFindMessage :: String}
+  | ReliabilityFailedToFindMsg
+      { missingMsgIndex :: Int
+      , sentMessagesLength :: Int
+      , knownAckForUs :: Int
+      , messageAckForUs :: Int
+      }
   | ReliabilityMissingPartyIndex {missingParty :: Party}
+  | ReceivedMalformedAcks
+      { fromParty :: Party
+      , partyAcks :: Vector Int
+      , numberOfParties :: Int
+      }
   deriving stock (Show, Eq, Generic)
   deriving anyclass (ToJSON, FromJSON)
 
@@ -123,6 +160,44 @@ instance Arbitrary (Vector Int) where
 
 instance Arbitrary ReliabilityLog where
   arbitrary = genericArbitrary
+
+-- | Handle for all persistence operations in the Reliability network layer.
+-- This handle takes care of storing and retreiving vector clock and all
+-- messages.
+data MessagePersistence m msg = MessagePersistence
+  { loadAcks :: m (Vector Int)
+  , saveAcks :: Vector Int -> m ()
+  , loadMessages :: m [Heartbeat msg]
+  , appendMessage :: Heartbeat msg -> m ()
+  }
+
+-- | Create 'MessagePersistence' out of 'PersistenceIncremental' and
+-- 'Persistence' handles. This handle loads and saves acks (vector clock data)
+-- and can load and append network messages.
+-- On start we construct empty ack vector from all parties in case nothing
+-- is stored on disk.
+-- NOTE: This handle is returned in the underlying context just for the sake of
+-- convenience.
+mkMessagePersistence ::
+  (MonadThrow m, FromJSON msg, ToJSON msg) =>
+  Int ->
+  PersistenceIncremental (Heartbeat msg) m ->
+  Persistence (Vector Int) m ->
+  MessagePersistence m msg
+mkMessagePersistence numberOfParties msgPersistence ackPersistence =
+  MessagePersistence
+    { loadAcks = do
+        macks <- load ackPersistence
+        case macks of
+          Nothing -> pure $ replicate numberOfParties 0
+          Just acks -> pure acks
+    , saveAcks = \acks -> do
+        save ackPersistence acks
+    , loadMessages = do
+        loadAll msgPersistence
+    , appendMessage = \msg -> do
+        append msgPersistence msg
+    }
 
 -- | Middleware function to handle message counters tracking and resending logic.
 --
@@ -136,6 +211,8 @@ withReliability ::
   (MonadThrow (STM m), MonadThrow m, MonadAsync m) =>
   -- | Tracer for logging messages.
   Tracer m ReliabilityLog ->
+  -- | Our persistence handle
+  MessagePersistence m msg ->
   -- | Our own party identifier.
   Party ->
   -- | Other parties' identifiers.
@@ -143,60 +220,68 @@ withReliability ::
   -- | Underlying network component providing consuming and sending channels.
   NetworkComponent m (Authenticated (ReliableMsg (Heartbeat msg))) (ReliableMsg (Heartbeat msg)) a ->
   NetworkComponent m (Authenticated (Heartbeat msg)) (Heartbeat msg) a
-withReliability tracer me otherParties withRawNetwork callback action = do
-  ackCounter <- newTVarIO $ replicate (length allParties) 0
-  sentMessages <- newTVarIO IMap.empty
+withReliability tracer MessagePersistence{saveAcks, loadAcks, appendMessage, loadMessages} me otherParties withRawNetwork callback action = do
+  acksCache <- loadAcks >>= newTVarIO
   resendQ <- newTQueueIO
   let ourIndex = fromMaybe (error "This cannot happen because we constructed the list with our party inside.") (findPartyIndex me)
   let resend = writeTQueue resendQ
-  withRawNetwork (reliableCallback ackCounter sentMessages resend ourIndex) $ \network@Network{broadcast} -> do
+  withRawNetwork (reliableCallback acksCache resend ourIndex) $ \network@Network{broadcast} -> do
     withAsync (forever $ atomically (readTQueue resendQ) >>= broadcast) $ \_ ->
-      reliableBroadcast ourIndex ackCounter sentMessages network
+      reliableBroadcast ourIndex acksCache network
  where
   allParties = fromList $ sort $ me : otherParties
-
-  reliableBroadcast ourIndex ackCounter sentMessages Network{broadcast} =
+  reliableBroadcast ourIndex acksCache Network{broadcast} =
     action $
       Network
         { broadcast = \msg ->
             case msg of
               Data{} -> do
-                ackCounter' <- atomically $ incrementCountersFor msg
-                traceWith tracer (BroadcastCounter ourIndex ackCounter')
-                broadcast $ ReliableMsg ackCounter' msg
+                newAckCounter <- incrementAckCounter
+                appendMessage msg
+                saveAcks newAckCounter
+                traceWith tracer (BroadcastCounter ourIndex newAckCounter)
+                broadcast $ ReliableMsg newAckCounter msg
               Ping{} -> do
-                acks <- readTVarIO ackCounter
+                acks <- readTVarIO acksCache
+                saveAcks acks
                 traceWith tracer (BroadcastPing ourIndex acks)
                 broadcast $ ReliableMsg acks msg
         }
    where
-    incrementCountersFor msg = do
-      acks <- readTVar ackCounter
-      writeTVar ackCounter $ constructAcks acks ourIndex
-      modifyTVar' sentMessages (insertNewMsg msg)
-      readTVar ackCounter
+    incrementAckCounter = atomically $ do
+      acks <- readTVar acksCache
+      let newAcks = constructAcks acks ourIndex
+      writeTVar acksCache newAcks
+      pure newAcks
 
-  reliableCallback ackCounter sentMessages resend ourIndex (Authenticated (ReliableMsg acks msg) party) = do
+  reliableCallback acksCache resend ourIndex (Authenticated (ReliableMsg acks msg) party) = do
     if length acks /= length allParties
-      then ignoreMalformedMessages
+      then
+        traceWith
+          tracer
+          ReceivedMalformedAcks
+            { fromParty = party
+            , partyAcks = acks
+            , numberOfParties = length allParties
+            }
       else do
         eShouldCallbackWithKnownAcks <- atomically $ runMaybeT $ do
+          loadedAcks <- lift $ readTVar acksCache
           partyIndex <- hoistMaybe $ findPartyIndex party
           messageAckForParty <- hoistMaybe (acks !? partyIndex)
-          knownAcks <- lift $ readTVar ackCounter
-          knownAckForParty <- hoistMaybe $ knownAcks !? partyIndex
+          knownAckForParty <- hoistMaybe $ loadedAcks !? partyIndex
           if
               | isPing msg ->
                   -- we do not update indices on Pings but we do propagate them
-                  return (True, partyIndex, knownAcks)
+                  return (True, partyIndex, loadedAcks)
               | messageAckForParty == knownAckForParty + 1 -> do
                   -- we update indices for next in line messages and propagate them
-                  let newAcks = constructAcks knownAcks partyIndex
-                  lift $ writeTVar ackCounter newAcks
+                  let newAcks = constructAcks loadedAcks partyIndex
+                  lift $ writeTVar acksCache newAcks
                   return (True, partyIndex, newAcks)
               | otherwise ->
                   -- other messages are dropped
-                  return (False, partyIndex, knownAcks)
+                  return (False, partyIndex, loadedAcks)
 
         case eShouldCallbackWithKnownAcks of
           Just (shouldCallback, partyIndex, knownAcks) -> do
@@ -207,50 +292,40 @@ withReliability tracer me otherParties withRawNetwork callback action = do
               else traceWith tracer (Ignored acks knownAcks partyIndex)
 
             when (isPing msg) $
-              resendMessagesIfLagging resend partyIndex sentMessages knownAcks acks ourIndex
+              resendMessagesIfLagging resend partyIndex knownAcks acks ourIndex
           Nothing -> pure ()
-
-  ignoreMalformedMessages = pure ()
 
   constructAcks acks wantedIndex =
     zipWith (\ack i -> if i == wantedIndex then ack + 1 else ack) acks partyIndexes
 
   partyIndexes = generate (length allParties) id
 
-  resendMessagesIfLagging resend partyIndex sentMessages knownAcks messageAcks myIndex = do
+  resendMessagesIfLagging resend partyIndex knownAcks messageAcks myIndex = do
     let mmessageAckForUs = messageAcks !? myIndex
     let mknownAckForUs = knownAcks !? myIndex
     case (mmessageAckForUs, mknownAckForUs) of
       (Just messageAckForUs, Just knownAckForUs) ->
-         -- We resend messages if our peer notified us that it's lagging behind our
-         -- latest message sent
-         when (messageAckForUs < knownAckForUs) $ do
-           let missing = fromList [messageAckForUs + 1 .. knownAckForUs]
-           messages <- readTVarIO sentMessages
-           forM_ missing $ \idx -> do
-             case messages IMap.!? idx of
-               Nothing ->
-                 traceWith tracer $
-                   ReliabilityFailedToFindMsg $
-                     "FIXME: this should never happen, there's no sent message at index "
-                       <> show idx
-                       <> ", messages length = "
-                       <> show (IMap.size messages)
-                       <> ", latest message ack: "
-                       <> show knownAckForUs
-                       <> ", acked: "
-                       <> show messageAckForUs
-               Just missingMsg -> do
-                 let newAcks' = zipWith (\ack i -> if i == myIndex then idx else ack) knownAcks partyIndexes
-                 traceWith tracer (Resending missing messageAcks newAcks' partyIndex)
-                 atomically $ resend $ ReliableMsg newAcks' missingMsg
+        -- We resend messages if our peer notified us that it's lagging behind our
+        -- latest message sent
+        when (messageAckForUs < knownAckForUs) $ do
+          let missing = fromList [messageAckForUs + 1 .. knownAckForUs]
+          storedMessages <- loadMessages
+          let messages = IMap.fromList (zip [1 ..] storedMessages)
+          forM_ missing $ \idx -> do
+            case messages IMap.!? idx of
+              Nothing ->
+                traceWith tracer $
+                  ReliabilityFailedToFindMsg
+                    { missingMsgIndex = idx
+                    , sentMessagesLength = IMap.size messages
+                    , knownAckForUs = knownAckForUs
+                    , messageAckForUs = messageAckForUs
+                    }
+              Just missingMsg -> do
+                let newAcks' = zipWith (\ack i -> if i == myIndex then idx else ack) knownAcks partyIndexes
+                traceWith tracer (Resending missing messageAcks newAcks' partyIndex)
+                atomically $ resend $ ReliableMsg newAcks' missingMsg
       _ -> pure ()
-
-  -- Find the maximum index and increment it by one to store the next message.
-  insertNewMsg msg m =
-    case IMap.lookupMax m of
-      Nothing -> IMap.insert 1 msg m
-      Just (k, _) -> IMap.insert (k + 1) msg m
 
   -- Find the index of a party in the list of all parties.
   -- NOTE: This should never fail.