+++ hydra-node/test/Hydra/Network/ReliabilitySpec.hs
@@ -2,7 +2,7 @@
 
 module Hydra.Network.ReliabilitySpec where
 
-import Hydra.Prelude hiding (empty, fromList, head, unlines)
+import Hydra.Prelude hiding (empty, fromList, head, replicate, unlines)
 import Test.Hydra.Prelude
 
 import Control.Concurrent.Class.MonadSTM (
@@ -15,12 +15,22 @@ import Control.Concurrent.Class.MonadSTM (
  )
 import Control.Monad.IOSim (runSimOrThrow)
 import Control.Tracer (Tracer (..), nullTracer)
-import Data.Vector (Vector, empty, fromList, head, snoc)
+import Data.Sequence.Strict ((|>))
+import Data.Vector (Vector, empty, fromList, head, replicate, snoc)
+import qualified Data.Vector as Vector
 import Hydra.Network (Network (..))
 import Hydra.Network.Authenticate (Authenticated (..))
 import Hydra.Network.Heartbeat (Heartbeat (..), withHeartbeat)
-import Hydra.Network.Reliability (ReliabilityLog (..), ReliableMsg (..), withReliability)
+import Hydra.Network.Reliability (MessagePersistence (..), ReliabilityLog (..), ReliableMsg (..), withReliability)
 import Hydra.Node.Network (withFlipHeartbeats)
+import Hydra.Persistence (
+  Persistence (..),
+  PersistenceIncremental (..),
+  createPersistence,
+  createPersistenceIncremental,
+ )
+import System.Directory (doesFileExist)
+import System.FilePath ((</>))
 import System.Random (mkStdGen, uniformR)
 import Test.Hydra.Fixture (alice, bob, carol)
 import Test.QuickCheck (
@@ -29,7 +39,8 @@ import Test.QuickCheck (
   counterexample,
   generate,
   tabulate,
-  (===), within,
+  within,
+  (===),
  )
 import Prelude (unlines)
 
@@ -88,11 +99,12 @@ spec = parallel $ do
     prop "broadcast messages to the network assigning a sequential id" $ \(messages :: [String]) ->
       let sentMsgs = runSimOrThrow $ do
             sentMessages <- newTVarIO empty
+            persistence <- mockMessagePersistence 1
 
-            withReliability nullTracer alice [] (captureOutgoing sentMessages) noop $ \Network{broadcast} -> do
+            withReliability nullTracer persistence alice [] (captureOutgoing sentMessages) noop $ \Network{broadcast} -> do
               mapM_ (broadcast . Data "node-1") messages
 
-            fromList . toList <$> readTVarIO sentMessages
+            fromList . Vector.toList <$> readTVarIO sentMessages
        in head . knownMessageIds <$> sentMsgs `shouldBe` fromList [1 .. (length messages)]
 
     -- this test is quite critical as it demonstrates messages dropped are properly managed and resent to the
@@ -107,14 +119,16 @@ spec = parallel $ do
             randomSeed <- newTVarIO $ mkStdGen seed
             aliceToBob <- newTQueueIO
             bobToAlice <- newTQueueIO
+            alicePersistence <- mockMessagePersistence 2
+            bobPersistence <- mockMessagePersistence 2
             let
               -- this is a NetworkComponent that broadcasts authenticated messages
               -- mediated through a read and a write TQueue but drops 0.2 % of them
               aliceFailingNetwork = failingNetwork randomSeed alice (bobToAlice, aliceToBob)
               bobFailingNetwork = failingNetwork randomSeed bob (aliceToBob, bobToAlice)
 
-              bobReliabilityStack = reliabilityStack bobFailingNetwork emittedTraces "bob" bob [alice]
-              aliceReliabilityStack = reliabilityStack aliceFailingNetwork emittedTraces "alice" alice [bob]
+              bobReliabilityStack = reliabilityStack bobPersistence bobFailingNetwork emittedTraces "bob" bob [alice]
+              aliceReliabilityStack = reliabilityStack alicePersistence aliceFailingNetwork emittedTraces "alice" alice [bob]
 
               runAlice = runPeer aliceReliabilityStack "alice" messagesReceivedByAlice messagesReceivedByBob aliceToBobMessages bobToAliceMessages
               runBob = runPeer bobReliabilityStack "bob" messagesReceivedByBob messagesReceivedByAlice bobToAliceMessages aliceToBobMessages
@@ -122,21 +136,24 @@ spec = parallel $ do
             concurrently_ runAlice runBob
 
             logs <- readTVarIO emittedTraces
-            aliceReceived <- toList <$> readTVarIO messagesReceivedByAlice
-            bobReceived <- toList <$> readTVarIO messagesReceivedByBob
+            aliceReceived <- Vector.toList <$> readTVarIO messagesReceivedByAlice
+            bobReceived <- Vector.toList <$> readTVarIO messagesReceivedByBob
             pure (aliceReceived, bobReceived, logs)
          in
-          within 1000000 $ msgReceivedByBob
-            === aliceToBobMessages
-            & counterexample (unlines $ show <$> reverse traces)
-            & tabulate "Messages from Alice to Bob" ["< " <> show ((length msgReceivedByBob `div` 10 + 1) * 10)]
-            & tabulate "Messages from Bob to Alice" ["< " <> show ((length msgReceivedByAlice `div` 10 + 1) * 10)]
+          within 1000000 $
+            msgReceivedByBob
+              === aliceToBobMessages
+              & counterexample (unlines $ show <$> reverse traces)
+              & tabulate "Messages from Alice to Bob" ["< " <> show ((length msgReceivedByBob `div` 10 + 1) * 10)]
+              & tabulate "Messages from Bob to Alice" ["< " <> show ((length msgReceivedByAlice `div` 10 + 1) * 10)]
 
     it "broadcast updates counter from peers" $ do
       let receivedMsgs = runSimOrThrow $ do
             sentMessages <- newTVarIO empty
+            alicePersistence <- mockMessagePersistence 2
             withReliability
               nullTracer
+              alicePersistence
               alice
               [bob]
               ( \incoming action -> do
@@ -148,9 +165,52 @@ spec = parallel $ do
               $ \Network{broadcast} -> do
                 threadDelay 1
                 broadcast (Data "node-1" msg)
-            toList <$> readTVarIO sentMessages
+            Vector.toList <$> readTVarIO sentMessages
 
       receivedMsgs `shouldBe` [ReliableMsg (fromList [1, 1]) (Data "node-1" msg)]
+
+    it "appends messages to disk and can load them back" $ do
+      withTempDir "network-messages-persistence" $ \tmpDir -> do
+        Persistence{load, save} <- createPersistence $ tmpDir <> "/acks"
+        PersistenceIncremental{loadAll, append} <- createPersistenceIncremental $ tmpDir <> "/network-messages"
+
+        let messagePersistence =
+              MessagePersistence
+                { loadAcks = do
+                    mloaded <- load
+                    case mloaded of
+                      Nothing -> pure $ replicate (length [alice, bob]) 0
+                      Just acks -> pure acks
+                , saveAcks = save
+                , loadMessages = loadAll
+                , appendMessage = append
+                }
+
+        receivedMsgs <- do
+          sentMessages <- newTVarIO empty
+          withReliability
+            nullTracer
+            messagePersistence
+            alice
+            [bob]
+            ( \incoming action -> do
+                concurrently_
+                  (action $ Network{broadcast = \m -> atomically $ modifyTVar' sentMessages (`snoc` m)})
+                  (incoming (Authenticated (ReliableMsg (fromList [0, 1]) (Data "node-2" msg)) bob))
+            )
+            noop
+            $ \Network{broadcast} -> do
+              threadDelay 1
+              broadcast (Data "node-1" msg)
+          Vector.toList <$> readTVarIO sentMessages
+
+        receivedMsgs `shouldBe` [ReliableMsg (fromList [1, 1]) (Data "node-1" msg)]
+
+        doesFileExist (tmpDir </> "network-messages") `shouldReturn` True
+        loadAll `shouldReturn` [Data "node-1" msg]
+
+        doesFileExist (tmpDir </> "acks") `shouldReturn` True
+        load `shouldReturn` Just (fromList [1, 1])
  where
   runPeer reliability partyName receivedMessageContainer sentMessageContainer messagesToSend expectedMessages =
     reliability (capturePayload receivedMessageContainer) $ \Network{broadcast} -> do
@@ -162,10 +222,10 @@ spec = parallel $ do
         (waitForAllMessages expectedMessages receivedMessageContainer)
         (waitForAllMessages messagesToSend sentMessageContainer)
 
-  reliabilityStack underlyingNetwork tracesContainer nodeId party peers =
+  reliabilityStack persistence underlyingNetwork tracesContainer nodeId party peers =
     withHeartbeat nodeId noop $
       withFlipHeartbeats $
-        withReliability (captureTraces tracesContainer) party peers underlyingNetwork
+        withReliability (captureTraces tracesContainer) persistence party peers underlyingNetwork
 
   failingNetwork seed peer (readQueue, writeQueue) callback action =
     withAsync
@@ -193,12 +253,14 @@ noop = const $ pure ()
 aliceReceivesMessages :: [Authenticated (ReliableMsg (Heartbeat msg))] -> [Authenticated (Heartbeat msg)]
 aliceReceivesMessages messages = runSimOrThrow $ do
   receivedMessages <- newTVarIO empty
+  alicePersistence <- mockMessagePersistence 3
 
   let baseNetwork incoming _ = mapM incoming messages
 
       aliceReliabilityStack =
         withReliability
           nullTracer
+          alicePersistence
           alice
           [bob, carol]
           baseNetwork
@@ -206,7 +268,7 @@ aliceReceivesMessages messages = runSimOrThrow $ do
   void $ aliceReliabilityStack (captureIncoming receivedMessages) $ \_action ->
     pure [()]
 
-  toList <$> readTVarIO receivedMessages
+  Vector.toList <$> readTVarIO receivedMessages
 
 captureIncoming :: MonadSTM m => TVar m (Vector p) -> p -> m ()
 captureIncoming receivedMessages msg =
@@ -229,3 +291,15 @@ captureTraces ::
   Tracer m ReliabilityLog
 captureTraces tvar = Tracer $ \msg -> do
   atomically $ modifyTVar' tvar (msg :)
+
+mockMessagePersistence :: Int -> MonadSTM m => m (MessagePersistence m msg)
+mockMessagePersistence numberOfParties = do
+  acks <- newTVarIO $ replicate numberOfParties 0
+  messages <- newTVarIO mempty
+  pure $
+    MessagePersistence
+      { loadAcks = readTVarIO acks
+      , saveAcks = atomically . writeTVar acks
+      , loadMessages = toList <$> readTVarIO messages
+      , appendMessage = \msg -> atomically $ modifyTVar' messages (|> msg)
+      }